Add galvan-hir crate with typed HIR definitions
Introduces a new workspace crate that will host Galvan's typed
high-level IR and typechecker:

- HIR nodes mirror the AST but reify control flow (if/for/try/else)
  as dedicated nodes instead of magic function calls, and every
  expression stores its inferred type and ownership
- Ownership and type coercions become explicit Adjustment values
  (borrow, mut-borrow, deref, to_owned, Some/Ok/Err wrapping, ...)
  so code generation no longer makes ownership decisions
- Moves the builtin type table (mapping/builtins) and the diagnostics
  infrastructure (ErrorCollector et al) from galvan-transpiler into
  galvan-hir so the upcoming typechecker can use them; the transpiler
  re-exports them from the new location

No behavior change.
