+++ Cargo.toml
@@ -11,6 +11,7 @@ galvan-files = { path = "galvan-files", version = "0.0.3" }
 galvan-test-macro = { path = "galvan-test-macro", version = "0.0.3" }
 galvan-ast-macro = { path = "galvan-ast-macro", version = "0.0.3"}
 galvan-resolver = { path = "galvan-resolver", version = "0.0.3" }
+galvan-hir = { path = "galvan-hir", version = "0.0.3" }
 galvan-into-ast = { path = "galvan-into-ast", version = "0.0.3" }
 galvan-parse = { path = "galvan-parse", version = "0.0.3" }
 
@@ -47,6 +48,7 @@ members = [
   # "galvan-core",
   "galvan-files",
   "galvan-resolver",
+  "galvan-hir",
   "galvan-transpiler",
   "galvan-test-macro",
   "galvan-test",
