+++ galvan-transpiler/Cargo.toml
@@ -21,6 +21,7 @@ itertools.workspace = true
 galvan-files.workspace = true
 galvan-ast.workspace = true
 galvan-resolver.workspace = true
+galvan-hir.workspace = true
 galvan-into-ast.workspace = true
 
 [dev-dependencies]
