+++ galvan-transpiler/src/context.rs
@@ -1,16 +1,13 @@
 use crate::mapping::Mapping;
-use galvan_ast::{
-    AstNode, EmptyTypeDecl, FnDecl, SegmentedAsts, ToplevelItem, TypeDecl, Visibility,
-    VisibilityKind,
-};
-use galvan_files::Source;
+use galvan_ast::SegmentedAsts;
 use galvan_resolver::{LookupContext, LookupError};
+
+pub use galvan_hir::builtins::predefined_from;
 
 #[derive(Debug, Default)]
 pub struct Context<'a> {
     pub lookup: LookupContext<'a>,
     pub mapping: Mapping,
-    // pub scope: Scope,
 }
 
 impl<'a> Context<'a> {
@@ -24,36 +21,5 @@ impl<'a> Context<'a> {
     pub fn with(mut self, asts: &'a SegmentedAsts) -> Result<Self, LookupError> {
         self.lookup = self.lookup.with(asts)?;
         Ok(self)
-    }
-}
-
-pub fn predefined_from(mapping: &Mapping, functions: Vec<FnDecl>) -> SegmentedAsts {
-    let types = mapping
-        .types
-        .keys()
-        .map(|ident| ToplevelItem {
-            item: TypeDecl::Empty(EmptyTypeDecl {
-                visibility: Visibility::new(VisibilityKind::Inherited, ident.span().clone()),
-                ident: ident.clone(),
-                span: ident.span().clone(),
-            }),
-            source: Source::Missing,
-        })
-        .collect();
-    let tests = vec![];
-    let functions = functions
-        .into_iter()
-        .map(|f| ToplevelItem {
-            item: f,
-            source: Source::Builtin,
-        })
-        .collect();
-    let main = None;
-    SegmentedAsts {
-        types,
-        functions,
-        tests,
-        main,
-        cmds: vec![],
     }
 }
